Add a wrapper for INIH reader to represent config
VHAT-228

Adding a Config class that allows to read config options from *.ini file.

This class is a simple wrapper for INIH reader, with minor modifications:
 - allow empty string as a value
 - return an int for integer types to satify Google style

Throw exception in wrapper constructor in case of file open or parse
errors. This excpetion is a placeholder for a better error handling in
future.