Add v4l impl for Streamer interface
VHAT-327

Add a v4l2 implementation of video streamer interface. The 3 channel
BGR matrix is returned as a frame.

RGB 4-byte format is hard-coded. Device and frame geometry can be
configured.

Frames are captured in a thread and each frame is copied to a dedicated
buffer. Thus a cv matrix can be safely created from this buffer.

In case of init errors, exception is thrown. In case of errors
encountered during capture, Frame() methods simply returns false.