Rename all CLI converters to ToCli and FromCli
VHAT-1794

For uniformality rename all converters so the name with demonstrate the
direction of conversion, rather then what types are converted.

To do so use the function of same name, and simply overload it.

Split converters "TO" and "FROM" into different files. Also separated
exception converter, as it has different semantics.